[refactor] Moving code inside `pkg` directory
