+++ main.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/3scale-labs/kamwiel/routing"
+import "github.com/3scale-labs/kamwiel/pkg/routing"
 
 func main() {
 	routing.Start()
